updated Airfield, CargoPlane, Fighter Jet, Jet, JetImpl, JetApplication

+++ src/com/skilldistillery/jets/JetImpl.java
@@ -6,7 +6,7 @@ public class JetImpl extends Jet{
 		
 	}
 
-	public JetImpl(String model, int speed, double range, long price) {
+	public JetImpl(String model, double speed, int range, long price) {
 		super(model, speed, range, price);
 		
 	}
